+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - switched to underscores in project name
+- `from-video-file` can process multiple files now
 
 
 0.0.1 (2024-05-06)